requirements.txt
streamlit
pandas
pydeck
Pillow
xlrd>=2.0.1
